renamed `legacy_app.py`
