+++ src/py/flwr/simulation/__init__.py
@@ -23,7 +23,7 @@ from flwr.simulation.simulationio_connection import SimulationIoConnection
 is_ray_installed = importlib.util.find_spec("ray") is not None
 
 if is_ray_installed:
-    from flwr.simulation.app import start_simulation
+    from flwr.simulation.legacy_app import start_simulation
 else:
     RAY_IMPORT_ERROR: str = """Unable to import module `ray`.
